==> 03-analysis-demographics.Rmd <==
---
title: "Analysis of demographics"
output:
  html_document:
    theme: lumen
    toc: true
    toc_float: false
    code_download: true
    highlight: tango
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "docs") })
---

```{r warning=FALSE, message=FALSE}
library(readr)
library(dplyr)
library(scales)
library(tidyr)
library(ggplot2)
library(cowplot)
library(stringr)
library(forcats)
library(ggallin)

theme_set(theme_bw())
```

## Read in data
```{r}
alpha_threshold <- qnorm(0.975)
countries <- c("France", "Germany", "Spain", "Italy", "Singapore", "USA" )
countries_code <- c("FR", "DE", "ES", "IT", "SG", "US")
colorBlindGrey6 <- c("#0072B2", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#CC79A7")
races <- c("black", "hispanic_latino", "white", "asian", 
           "hawaiian_pacific_islander", "american_indian")  
Races <- rev(c('White', 'Black', 'Hispanic/Latino', 'Asian', 
               'American Indian', 'Hawaiian/Pacific Islander', 'Other/Unknown'))

demo_desc <- read_csv('results/demo_prelim.csv') %>% 
  filter(siteid != 'SITE157', siteid != 'SITE712') %>% 
  mutate(Country = fct_relevel(Country, countries))

demo_perc <- demo_desc %>% 
  mutate_at(vars(- num_patients_all, - siteid, - Country, - median_age),
            list(~ . / .data$num_patients_all)) %>% 
  mutate(Country = as.factor(Country)) 

demo_perc %>% 
  rename('perc_patients_ever_severe' = num_patients_ever_severe,
         'perc_patients_never_severe' = num_patients_never_severe) %>% 
  write_csv('results/demographics.csv')
```

## Overall statistics
```{r}
demo_desc %>% 
  summarise(across(c(num_patients_all, female, male, unknown_other_sex), sum)) 

demo_desc %>% 
  summarise(across(c(num_patients_all, races, unknown_other_race), sum)) %>% 
  print()

unknowns <- demo_perc %>%
  select(siteid, Country, contains('unknown')) %>% 
  pivot_longer(- c(siteid, Country))
unknowns %>% filter(value < 0)
unknowns %>% filter(value < 1, value > 0.1)

# unknowns %>% 
#   ggplot(aes(y = value, x = name)) +
#   ggbeeswarm::geom_beeswarm() +
#   facet_wrap(~ name, scales = 'free_x')
```

## Severity vs. gender/age/race?
```{r}
demo_gender <- demo_perc %>% filter(unknown_other_sex < 1)
lm(num_patients_ever_severe ~ female, demo_gender) %>% 
  summary()

age_perc <- demo_perc %>% 
  filter(unknown_other_age_group < 1) %>% 
  select(Country, num_patients_all, num_patients_ever_severe, contains('to'), ends_with('plus'))

lm(num_patients_ever_severe ~ . - Country - num_patients_all, age_perc) %>% 
  summary()

lm(num_patients_ever_severe ~ black + white + asian + hispanic_latino + 
     american_indian + hawaiian_pacific_islander, 
   data = filter(demo_perc, unknown_other_race < 1)) %>% 
  summary()

lm(num_patients_ever_severe ~ median_age, demo_perc) %>% 
  summary()
```

## Overall plot (A)

```{r}
gender_count <- demo_desc %>% 
  # filter(unknown_other_sex/num_patients_all < 1) %>% 
  # mutate(siteid = as.factor(siteid)) %>% 
  # mutate(num_patients_all_gender = female + male) %>% 
  group_by(Country) %>% 
  summarise(num_patients_country = sum(num_patients_all),
            male_count = sum(male),
            female_count = sum(female),
            .groups = 'drop') %>% 
  {.}

showtext::showtext_auto()
female <- intToUtf8(9792)
male <- intToUtf8(9794)

overall_plot <- gender_count %>% 
  mutate(male_count = -male_count) %>% 
  pivot_longer(- c(Country, num_patients_country)) %>% 
  ggplot(aes(y = fct_relevel(Country, rev(countries)), 
             x = value, color = name, 
             fill = Country)) +
  geom_col(position = 'identity') +
  scale_color_manual(values = c('white', 'white')) +
  scale_x_continuous(breaks = c(-5000, -2000, 0, 2000, 5000),
                     labels = c(5000, 2000, 0, 2000, 5000),
                     trans = ssqrt_trans) +
  geom_vline(xintercept = 0, color = 'white') +
  geom_text(data = . %>% filter(name == 'male_count'),
            aes(y = Country, x = value - 22*sqrt(abs(value)), 
                label = abs(value)), color = 'grey40', size = 3) +
    geom_text(data = . %>% filter(name == 'female_count'),
            aes(y = Country, x = value + 19*sqrt(abs(value)), 
                label = abs(value)), color = 'grey40', size = 3) +
  geom_text(data = filter(gender_count, Country == 'France'),
            aes(y = 6, x = female_count - 600), fontface = 2,
            label = female, size = 3, color = 'white') +
  geom_text(data = filter(gender_count, Country == 'France'),
            aes(y = 6, x = - male_count + 600), fontface = 2,
            label = male, size = 3, color = 'white') +
  labs(y = NULL, 
       x = 'Number of male/female patients per country', 
       fill = NULL) +
  scale_fill_manual(values = colorBlindGrey6) +
  theme(legend.position = 'None',
        plot.margin = margin(15, 5.5, 15.5, 55.5),
        axis.ticks.y = element_blank(),
        panel.grid.major.y = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        panel.spacing.y = unit(0, 'pt'),
        strip.background = element_blank(),
        strip.placement = 'inside',
        legend.background = element_blank(),
        legend.key.height = unit(1, 'line'),
        legend.title = element_text(hjust = 0, size = 9)) +
  NULL
```



## Age plot (B)

```{r}
age_count <- demo_desc %>% 
  filter(unknown_other_age_group/num_patients_all < 1) %>% 
  select(Country, num_patients_all, contains('to'), ends_with('plus')) %>% 
  group_by(Country) %>% 
  summarise(across(.fns = sum), .groups = 'drop') %>% 
  mutate_at(vars(- num_patients_all, - Country),
            list(~ . / .data$num_patients_all)) %>% 
  select(-num_patients_all) %>% 
  pivot_longer(- Country) %>% 
  mutate(name = case_when(
    name == '00to02' ~ '0to2',
    name == '03to05' ~ '3to5',
    name == '06to11' ~ '6to11',
    TRUE ~ name
  )) %>% 
  mutate(name = gsub('to', '–', name) %>% gsub('plus', '+', .))
age_count$name <- factor(age_count$name, levels = unique(age_count$name))

age_plot <- age_count  %>% 
  ggplot(aes(name, value, fill = Country, group = Country)) +
  geom_col(position = 'dodge2') + 
  coord_flip() +
  facet_grid(rows = vars(Country)) +
  scale_y_continuous(labels = scales::percent_format(accuracy = 1),
                     breaks = seq(0, 1, 0.2),
                     expand = expansion(add = c(0,0.02))) +
  scale_size_continuous(guide = FALSE) +
  labs(x = NULL, y = 'Patients proportion per age group') +
  scale_color_manual(values = colorBlindGrey6, drop = FALSE) +
  scale_fill_manual(values = colorBlindGrey6, drop = FALSE) +
  theme(legend.position = 'None',
        axis.text.y = element_text(size = 7),
        panel.grid.major.y = element_blank(),
        plot.margin = margin(15, 5.5, 5.5, 5.5, 'pt')) +
  NULL
```

## Severity vs. age plot (C)

```{r}
country_codes <- read_csv('https://raw.githubusercontent.com/lukes/ISO-3166-Countries-with-Regional-Codes/master/all/all.csv') %>%
  mutate(name = ifelse(name == 'United States of America', 'USA', name)) %>%
  filter(name %in% countries) %>%
  select(Country = name, coun_code = `alpha-2`) %>%
  mutate(coun_code = as.factor(coun_code) %>% fct_relevel(countries_code))

age_severity <- demo_perc %>% 
  left_join(country_codes, by = 'Country') %>%
  ggplot(aes(median_age, num_patients_ever_severe, 
             color = coun_code)) +
  geom_point(aes(size = num_patients_all), alpha = 0.8) +
  scale_y_continuous(labels = scales::percent_format(),
                     breaks = seq(0, 1, 0.2)) +
  labs(x = 'Median age estimate', 
       y = 'Severe case proportion', 
       color = NULL) +
  scale_size(trans = "log10", guide = FALSE) +
  scale_color_manual(values = colorBlindGrey6) +
  coord_cartesian(ylim = c(0, 1)) +
  theme(
    # legend.position = c(0.9, 0.7),
    legend.position = 'left',
    legend.background = element_blank(),
    plot.margin = margin(5.5, 5.5, 10.5, 15.5),
    panel.grid.minor = element_blank(),
    legend.key.height = unit(0.7, 'line'),
    legend.key.width = unit(0.4, 'line'),
    legend.title = element_text(hjust = 0, size = 9)) +
  NULL
```

## Race plot (D)

```{r}
race_count <- demo_desc %>% 
  filter(unknown_other_race/num_patients_all < 1) %>% 
  select(Country, num_patients_all, all_of(races), unknown_other_race) %>% 
  group_by(Country) %>% 
  summarise(across(.fns = sum), .groups = 'drop') %>% 
  mutate_at(vars(- num_patients_all, - Country),
            list(~ . / .data$num_patients_all)) %>% 
  select(-num_patients_all) %>% 
  pivot_longer(- Country) %>% 
  mutate(name = gsub('_', ' ', name) %>% 
           str_to_title(.) %>% 
           gsub('Hispanic ', 'Hispanic/', .) %>% 
           gsub('Hawaiian ', 'Hawaiian/', .) %>% 
           gsub('Unknown Other Race', 'Other/Unknown',. ),
         name = fct_relevel(name, Races))

race_plot <- race_count %>% 
  # remove one site in Italy that reported all white ethnicity
  filter(Country != 'Italy') %>% 
  ggplot(aes(name, value, fill = Country, group = Country)) +
  geom_col(position = 'dodge2') + 
  coord_flip() +
  facet_grid(rows = vars(Country)) +
  scale_y_continuous(labels = scales::percent_format(accuracy = 1),
                     breaks = seq(0, 1, 0.2),
                     expand = expansion(add = c(0,0.02))) +
  scale_size_continuous(guide = FALSE) +
  labs(x = NULL, y = 'Patient proportion per race/ethnicity group') +
  scale_color_manual(values = colorBlindGrey6, drop = FALSE) +
  scale_fill_manual(values = colorBlindGrey6, drop = FALSE) +
  theme(legend.position = 'None',
        axis.text.y = element_text(size = 8),
        panel.grid.major.y = element_blank(),
        axis.title.x = element_text(hjust = .8)) +
  NULL
```

## Composite plot

```{r fig.width=10, fig.height=8, warning=FALSE}
gender_race <- plot_grid(overall_plot, age_severity, race_plot, ncol = 1,
          rel_heights = c(0.7, 1, 1), 
          labels = c('A', 'B', 'C'), axis = 'r')
abcd <- plot_grid(gender_race, age_plot, ncol = 2,
                  rel_widths = c(1.25, 1 ), labels = c('', 'D'),
                  axis = 'b')
abcd
ggsave('figs/composite_plot.png', abcd, height = 6.5, width = 9)
ggsave('figs/tiffs/fig_2.tiff', abcd, height = 6.5, width = 9, dpi = 300)
```




```{r eval=FALSE, include=FALSE}
# gender_plot <- demo_gender %>% 
#   ggplot(aes(female, num_patients_ever_severe, 
#              color = Country)) +
#   geom_vline(xintercept = 0.5, linetype = 'dashed') +
#   geom_point(aes(size = num_patients_all), alpha = 0.8) +
#   scale_x_continuous(labels = scales::percent_format(),
#                      limits = c(0, 0.65),
#                      expand = expansion(add = c(0.18, 0))) +
#   scale_y_continuous(labels = scales::percent_format(),
#                      breaks = seq(0, 1, 0.2)) +
#   labs(
#     x = 'Proportion of female patients', 
#     y = 'Proportion of severe cases', 
#     color = NULL, 
#     size = 'Total number of patients
# at each site') +
#   scale_size(trans = "log10", guide = FALSE) +
#   scale_color_manual(values = colorBlindGrey6) +
#   theme(legend.position = c(0.15, 0.28),
#         plot.margin = margin(10.5, 5.5, 5.5, 45.5, 'pt'),
#         legend.background = element_blank(),
#         legend.key.height = unit(1, 'line'),
#         legend.title = element_text(hjust = 0, size = 9)) +
#   NULL

# age_radar <- age_perc %>% 
#   select(-contains('num')) %>% 
#   group_by(Country) %>% 
#   summarise(across(.fns = mean_se), .groups = 'drop') %>% 
#   mutate(Country = as.character(Country)) %>% 
#   rename(group = Country)
# 
# mtcars %>%
#      add_rownames( var = "group" ) %>%
#      mutate_each(funs(rescale), -group) %>%
#      tail(4) %>% select(1:10) -> mtcars_radar
# 
# ggradar(age_radar, 
#         values.radar = c(NA, paste0(c(25, 50), '%')),
#         group.colours = colorBlindGrey6[-4],
#         grid.mid = 0.25,
#         grid.max = 0.5,
#         base.size = 1)

# age_perc %>% 
#   pivot_longer(- c(Country, num_patients_all, num_patients_ever_severe)) %>% 
#   ggplot(aes(value, num_patients_ever_severe, 
#              color = Country)) +
#   facet_wrap(~ name) +
#   geom_point(aes(size = log2(num_patients_all)), alpha = 0.8) +
#   scale_x_continuous(labels = scales::percent_format(accuracy = 1)) +
#   scale_y_continuous(labels = scales::percent_format(accuracy = 1)) +
#   scale_size_continuous(guide = FALSE) +
#   labs(x = '% patients in each age group', y = '% severe cases') +
#   scale_color_manual(values = colorBlindGrey6, drop = FALSE) +
#   NULL

# age_perc %>% 
#   pivot_longer(- c(Country, num_patients_all, num_patients_ever_severe)) %>% 
#   ggplot(aes(name, value, color = Country)) +
#   # facet_grid(cols = vars(name), scales = 'free_x') +
#   geom_boxplot(alpha = 0.8) +
#   # geom_violin(aes(color = Country)) +
#   # geom_jitter() +
#   # geom_point(aes(size = log2(num_patients_all)), alpha = 0.8) +
#   # scale_x_continuous(labels = scales::percent_format(accuracy = 1)) +
#   # scale_y_continuous(labels = scales::percent_format(accuracy = 1)) +
#   scale_size_continuous(guide = FALSE) +
#   labs(x = '% patients in each age group', y = '% patient') +
#   scale_color_manual(values = colorBlindGrey6, drop = FALSE) +
#   scale_fill_manual(values = colorBlindGrey6, drop = FALSE) +
#   coord_polar() +
#   NULL
# 
# race_perc %>% 
#   pivot_longer(- c(Country, num_patients_all, num_patients_ever_severe)) %>% 
#   ggplot(aes(value, num_patients_ever_severe, 
#              color = Country)) +
#   facet_wrap(~ name) +
#   geom_point(aes(size = log2(num_patients_all)), alpha = 0.8) +
#   scale_x_continuous(labels = scales::percent_format(accuracy = 0.1)) +
#   scale_y_continuous(labels = scales::percent_format(accuracy = 0.1)) +
#   scale_size_continuous(guide = FALSE) +
#   labs(x = '% patients in each race/ethnicity group', y = '% severe cases') +
#   scale_color_manual(values = colorBlindGrey6, drop = FALSE, guide = FALSE) +
#   NULL
# 
# race_perc %>% 
#   pivot_longer(- c(Country, num_patients_all, num_patients_ever_severe)) %>% 
#   ggplot(aes(name, value, color = Country)) +
#   # facet_wrap(~ name) +
#   # geom_point(aes(size = log2(num_patients_all)), alpha = 0.8) +
#   geom_boxplot() +
#   # scale_x_continuous(labels = scales::percent_format(accuracy = 0.1)) +
#   scale_y_continuous(labels = scales::percent_format(accuracy = 0.1)) +
#   scale_size_continuous(guide = FALSE) +
#   labs(x = '% patients in each race/ethnicity group', y = '% severe cases') +
#   scale_color_manual(values = colorBlindGrey6, drop = FALSE, guide = FALSE) +
#   NULL
```
